Refactor plotting utilities and integrate into modules
